requirements.txt
py-cord
python-dotenv
PSNAWP
discord-py